Move SQL query function from main module
