+++ src/messageparser.py
@@ -1,19 +1,4 @@
 import argparse
-import logging
-
-
-def run_query(query_string, con):
-    """
-    Run a query on a SQLite database.
-    :param query_string: Query to run.
-    :param con: Connection to run query on.
-    :return: Result of query.
-    """
-    logging.info("Running query: '{}'".format(query_string))
-    cur = con.cursor()
-    cur.execute(query_string)
-    logging.info("Query ran successfully.")
-    return cur.fetchall()
 
 
 def parse_arguments():
